exemple-sc/pyproject.toml
[tool.poetry]
name = "exemple-sc"
version = "0.1.0"
description = ""
authors = ["Vincent Perrollaz <vincent.perrollaz@laposte.net>"]
readme = "README.md"
packages = [{include = "exemple_sc"}]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.1"
typer = "^0.9.0"
pyserde = "^0.14.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
black = "^24.2.0"
mypy = "^1.8.0"
pytest = "^8.0.2"
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
